Move small test files to the 'tests' directory
The tests files are cluttering up my Django app! No good.

In some future scenario, I might move more of the top level files
to a directory, but for sure the tests files can live elsewhere.

The main tests file, nofo_test.py has not been moved because I feel
like it should be refactored rather than just moved outright (eg,
see the last commit).
